+++ src/main/java/vendingmachine/model/vo/Money.java
@@ -1,4 +1,4 @@
-package vendingmachine.model.money;
+package vendingmachine.model.vo;
 
 import static vendingmachine.validation.NumberValidator.isNotMultipleOfTen;
 import static vendingmachine.validation.NumberValidator.isNotPositiveInteger;
